better types, id validation

+++ src/actions/actions.ts
@@ -1,13 +1,11 @@
 "use server";
 import prisma from "@/lib/db";
-import { PetEssentials } from "@/lib/types";
 import { sleep } from "@/lib/utils";
-import { petFormSchema } from "@/lib/validations";
-import { Pet } from "@prisma/client";
+import { petFormSchema, petIdSchema } from "@/lib/validations";
 import { revalidatePath } from "next/cache";
 
-export async function addPet(pet: PetEssentials) {
-  await sleep(2000);
+export async function addPet(pet: unknown) {
+  await sleep(1000);
 
   const validatedPet = petFormSchema.safeParse(pet);
 
@@ -27,12 +25,27 @@ export async function addPet(pet: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function updatePet(petId: Pet["id"], newPetData: PetEssentials) {
+export async function updatePet(petId: unknown, newPetData: unknown) {
   await sleep(1000);
+
+  const validatedPet = petFormSchema.safeParse(newPetData);
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPet.success) {
+    return {
+      message: "server side validation failed: newPetData",
+    };
+  }
+  if (!validatedPetId.success) {
+    return {
+      message: "server side validation failed: petId",
+    };
+  }
+
   try {
     await prisma.pet.update({
-      where: { id: petId },
-      data: newPetData,
+      where: { id: validatedPetId.data },
+      data: validatedPet.data,
     });
   } catch (error) {
     return { message: "Error updating pet: " + error };
@@ -41,11 +54,20 @@ export async function updatePet(petId: Pet["id"], newPetData: PetEssentials) {
   revalidatePath("/app", "layout");
 }
 
-export async function deletePet(petId: Pet["id"]) {
+export async function deletePet(petId: unknown) {
   await sleep(1000);
+
+  const validatedPetId = petIdSchema.safeParse(petId);
+
+  if (!validatedPetId.success) {
+    return {
+      message: "server side validation failed: petId",
+    };
+  }
+
   try {
     await prisma.pet.delete({
-      where: { id: petId },
+      where: { id: validatedPetId.data },
     });
   } catch (error) {
     return { message: "Error deleting pet: " + error };
